// app/UI/components/block/types.ts
import {COLORS} from '@theme/colors';
import {ViewProps, FlexAlignType, ViewStyle} from 'react-native';

export interface IBlock extends ViewProps {
  flex?: boolean;
  flexShrink?: boolean;
  flexGrow?: boolean;
  row?: boolean;
  column?: boolean;
  shadow?: boolean;
  backgroundColor?: string;
  space?: 'between' | 'around' | 'evenly';
  padding?: number;
  margin?: number;
  alignStart?: boolean;
  alignCenter?: boolean;
  alignEnd?: boolean;
  wrap?: boolean;
  justifyCenter?: boolean;
  justifyEnd?: boolean;
  justifyStart?: boolean;
  justifySpaceBetween?: boolean;
  paddingTop?: number;
  paddingBottom?: number;
  paddingLeft?: number;
  paddingRight?: number;
  marginBottom?: number;
  marginLeft?: number;
  marginRight?: number;
  marginTop?: number;
  paddingVertical?: number;
  paddingHorizontal?: number;
  marginVertical?: number;
  marginHorizontal?: number;
  radius?: number;
  height?: number;
  width?: number;
  square?: number;
  round?: number;
  borderWidth?: number;
  relative?: boolean;
  absolute?: boolean;
  top?: number;
  left?: number;
  right?: number;
  bottom?: number;
  borderColor?: keyof typeof COLORS;
  animated?: boolean;
  overflow?: 'hidden' | 'scroll' | 'visible';
  alignSelf?: 'auto' | FlexAlignType;
  style?: ViewStyle | ViewStyle[];
  opacity?: number;
  borderTopLeftRadius?: number;
  borderTopRightRadius?: number;
}